Translate Japanese.java
It's not the most formal of the Japanese launguage, but it's mostly formal with a pinch of friendliness.
